extract token cache routes and logic
