+++ logic/cacheLogic.js
@@ -20,9 +20,6 @@ const searchOptions = {
   keys: ['title', 'chainName', 'sender', 'preview.description', 'preview.title', 'url', 'topics']
 }
 
-// TODO route to index token and add to registry
-// TODO index true/false balance for accounts of tokens
-
 module.exports = class CacheLogic {
 
   constructor() {
@@ -70,17 +67,6 @@ module.exports = class CacheLogic {
     AsyncTipGeneratorsLogic.triggerGetTokenContractIndex(tips);
 
     return tips;
-  }
-
-  static async fetchTokenInfos() {
-    const fetchData = async () => {
-      const tips = await aeternity.getTips();
-      return AsyncTipGeneratorsLogic.triggerGetTokenContractIndex(tips);
-    };
-
-    return this.cache
-      ? this.cache.getOrSet(['fetchTokenInfos'], () => fetchData(), this.cache.shortCacheTime)
-      : fetchData();
   }
 
   static fetchChainNames() {
@@ -248,11 +234,6 @@ module.exports = class CacheLogic {
     }
 
     res.send(tips);
-  }
-
-  static async deliverTokenInfo(req, res) {
-    const tokenInfo = await CacheLogic.fetchTokenInfos();
-    res.send(tokenInfo);
   }
 
   static async deliverContractEvents(req, res) {
