+++ routes/cacheRoutes.js
@@ -24,8 +24,6 @@ router.get('/topics', CacheLogic.deliverTipTopics);
 
 router.get('/events', CacheLogic.deliverContractEvents);
 
-router.get('/tokenInfo', CacheLogic.deliverTokenInfo);
-
 router.get('/invalidate/tips', CacheLogic.invalidateTips);
 
 router.get('/invalidate/oracle', CacheLogic.invalidateOracle);
