+++ routes/tokenCacheRoutes.js
@@ -1,0 +1,12 @@
+const TokenCacheLogic = require('../logic/tokenCacheLogic.js');
+const Router = require('express').Router;
+
+const router = new Router();
+
+new TokenCacheLogic(); //calls init
+
+// Open api routes
+router.get('/tokenInfo', TokenCacheLogic.deliverTokenInfo);
+
+module.exports = router;
+
